Updated filter usage

+++ include/private/plugins/filter.h
@@ -125,7 +125,7 @@ namespace lsp
 
             protected:
                 static inline dspu::equalizer_mode_t get_eq_mode(ssize_t mode);
-                static inline void  decode_filter(size_t *ftype, size_t *slope, size_t mode);
+                static inline void  decode_filter(uint32_t *ftype, uint32_t *slope, size_t mode);
                 static size_t       decode_slope(size_t slope);
                 static bool         filter_has_width(size_t type);
                 static inline bool  adjust_gain(size_t filter_type);
